setting up cmd line arg, code skeleton etc
