+++ run.py
@@ -1,4 +1,4 @@
-#!/usr/bin/python
+#!/usr/bin/env python
 ## @package dnn-visualization
 #  @author Attila Borcs
 #
@@ -22,79 +22,50 @@ import tensorflow as tf
 import tensorflow.contrib.slim as slim
 from tensorflow.examples.tutorials.mnist import input_data
 import math
+import sys
+import getopt
 
-mnist = input_data.read_data_sets("MNIST_data/", one_hot=True)
+def usage():
+     print 'run.py -option'
+     print '<option>: -f <Fetch Data from web...>'
+     print '<option>: -w <Test on Webam...>'
+     print '<option>: -m <Test on MNIST Data...>'
 
-tf.reset_default_graph()
+def main():
+    try:
+        opts, args = getopt.getopt(sys.argv[1:], "fhw", ["fetch", "help", "webcam"])
+    except getopt.GetoptError as err:
 
-x = tf.placeholder(tf.float32, [None, 784], name="x-in")
-true_y = tf.placeholder(tf.float32, [None, 10], name="y-in")
-keep_prob = tf.placeholder("float")
+        print('<option is not recognized>')
+        usage()
+        sys.exit(2)
 
-x_image = tf.reshape(x, [-1, 28, 28, 1])
-hidden_1 = slim.conv2d(x_image, 5, [5, 5])
-pool_1 = slim.max_pool2d(hidden_1, [2, 2])
-hidden_2 = slim.conv2d(pool_1, 5, [5, 5])
-pool_2 = slim.max_pool2d(hidden_2, [2, 2])
-hidden_3 = slim.conv2d(pool_2, 20, [5, 5])
-hidden_3 = slim.dropout(hidden_3, keep_prob)
-out_y = slim.fully_connected(
-    slim.flatten(hidden_3), 10, activation_fn=tf.nn.softmax)
-
-cross_entropy = -tf.reduce_sum(true_y * tf.log(out_y))
-correct_prediction = tf.equal(tf.argmax(out_y, 1), tf.argmax(true_y, 1))
-accuracy = tf.reduce_mean(tf.cast(correct_prediction, "float"))
-train_step = tf.train.AdamOptimizer(1e-4).minimize(cross_entropy)
-
-batchSize = 50
-sess = tf.Session()
-init = tf.global_variables_initializer()
-sess.run(init)
-for i in range(1001):
-    batch = mnist.train.next_batch(batchSize)
-    sess.run(
-        train_step, feed_dict={x: batch[0],
-                               true_y: batch[1],
-                               keep_prob: 0.5})
-    if i % 100 == 0 and i != 0:
-        trainAccuracy = sess.run(
-            accuracy, feed_dict={x: batch[0],
-                                 true_y: batch[1],
-                                 keep_prob: 1.0})
-        print("step %d, training accuracy %g" % (i, trainAccuracy))
-
-testAccuracy = sess.run(
-    accuracy,
-    feed_dict={x: mnist.test.images,
-               true_y: mnist.test.labels,
-               keep_prob: 1.0})
-print("test accuracy %g" % (testAccuracy))
+    parse_data = False
+    for o, a in opts:
+        if o in ("-f", "--fetch"):
+            parse_data = True
+            print('Parsing Data...')
+        elif o in ("-h", "--help"):
+            usage()
+            sys.exit()
+        elif o in ("-w", "--webcam"):
+            print('Test on Webam...')
+        else:
+            assert False, "unhandled option"
+            usage()
 
 
-def getActivations(layer, stimuli):
-    units = sess.run(
-        layer,
-        feed_dict={x: np.reshape(stimuli, [1, 784], order='F'),
-                   keep_prob: 1.0})
-    plotNNFilter(units)
+if __name__ == "__main__":
+    main()
 
+#if __name__ == "__main__":
 
-def plotNNFilter(units):
-    filters = units.shape[3]
-    plt.figure(1, figsize=(20, 20))
-    n_columns = 6
-    n_rows = math.ceil(filters / n_columns) + 1
-    for i in range(filters):
-        plt.subplot(n_rows, n_columns, i + 1)
-        plt.title('Filter ' + str(i))
-        plt.imshow(units[0, :, :, i], interpolation="nearest", cmap="gray")
+#TODO: plug webcam images here
+#imageToUse = mnist.test.images[0]
 
-        imageToUse = mnist.test.images[0]
+#plt.imshow(
+#    np.reshape(imageToUse, [28, 28]), interpolation="nearest", cmap="gray")
 
-
-plt.imshow(
-    np.reshape(imageToUse, [28, 28]), interpolation="nearest", cmap="gray")
-
-getActivations(hidden_1, imageToUse)
-getActivations(hidden_2, imageToUse)
-getActivations(hidden_3, imageToUse)
+#getActivations(hidden_1, imageToUse)
+#getActivations(hidden_2, imageToUse)
+#getActivations(hidden_3, imageToUse)
